Moved some of the shared acceptance code into a util file.
