+++ .gitignore
@@ -13,6 +13,7 @@
 node_modules
 .nyc_output/
 coverage/
+/controller/aws*.sh
 testmerge.json
 .env.local*
 .env.development.local*
